Pull page-navigation from TopList to PageNavigation component + make it's final version #6
