+++ src/styles/page-navigation.scss
@@ -1,10 +1,9 @@
 .page-navigation {
-  width: 400px;
+  width: 200px;
   margin: 0 auto;
   display: flex;
   justify-content: space-between;
   align-items: center;
-  list-style-type: none;
 
   &_place_top-list {
     margin-top: 70px;
@@ -15,27 +14,21 @@
     text-decoration: none;
     background-size: 20px;
     background-position: center center;
+    padding: 10px;
 
     &_dir_next {
       background-image: url('../img/next-page.svg');
       background-repeat: no-repeat;
-      background-position: center center;
-      padding: 10px;
     }
 
     &_dir_prev {
       background-image: url('../img/previous-page.svg');
       background-repeat: no-repeat;
-      background-position: center center;
-      padding: 10px;
     }
+  }
 
-    &-number {
-      padding: 5px;
-    }
-
-    &:hover {
-      cursor: pointer;
-    }
+  &__info {
+    font-weight: 400;
+    font-size: 20px;
   }
 }
